// src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        Attachment: {
            /** @description Alternate text that describes what is in the media attachment, to be used for the visually impaired or when media attachments do not load */
            description?: string | null;
            /** @description The ID of the attachment in the database. */
            id?: string;
            pleroma?: {
                /** @description mime type of the attachment */
                mime_type?: string;
            };
            /**
             * Format: uri
             * @description The location of a scaled-down preview of the attachment
             */
            preview_url?: string;
            /**
             * Format: uri
             * @description The location of the full-size original attachment on the remote website. String (URL), or null if the attachment is local
             */
            remote_url?: string | null;
            /**
             * Format: uri
             * @description A shorter URL for the attachment
             */
            text_url?: string;
            /**
             * @description The type of the attachment
             * @enum {string}
             */
            type?: "image" | "video" | "audio" | "unknown";
            /**
             * Format: uri
             * @description The location of the original full-size attachment
             */
            url?: string;
        };
// ...
    };
    responses: never;
    parameters: {
        /**
         * @description Account ID or nickname
         * @example 123
         */
        accountIdOrNickname: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}